# Optimized set::erase