```cpp
template<int N, int I>
using mk_list = brigand::erase<
  brigand::wrap<brigand::range<int, 0, N>, brigand::set>,
  std::integral_constant<int, I>
>;
```
---

```cpp
using list = mk_list<SIZE, 0>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.08s - 37M | 00.09s - 39M | 00.14s - 45M | 00.28s - 66M
gcc | 00.05s - 21M | 00.06s - 23M | 00.08s - 26M | 00.13s - 34M | 00.28s - 61M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.09s - 39M | 00.13s - 43M | 00.20s - 51M
gcc | 00.05s - 21M | 00.07s - 22M | 00.08s - 25M | 00.13s - 30M | 00.21s - 41M

---

```cpp
using list = mk_list<SIZE, SIZE/4>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.08s - 37M | 00.10s - 40M | 00.14s - 46M | 00.29s - 68M
gcc | 00.05s - 20M | 00.06s - 23M | 00.08s - 26M | 00.13s - 35M | 00.29s - 63M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.09s - 39M | 00.13s - 43M | 00.20s - 51M
gcc | 00.05s - 21M | 00.06s - 22M | 00.08s - 25M | 00.13s - 30M | 00.21s - 41M

---

```cpp
using list = mk_list<SIZE, SIZE/2>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.10s - 40M | 00.15s - 46M | 00.30s - 69M
gcc | 00.05s - 20M | 00.06s - 23M | 00.08s - 26M | 00.14s - 36M | 00.30s - 65M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.09s - 39M | 00.13s - 43M | 00.20s - 51M
gcc | 00.05s - 21M | 00.07s - 22M | 00.08s - 25M | 00.13s - 30M | 00.21s - 41M

---

```cpp
using list = mk_list<SIZE, SIZE-1>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.10s - 40M | 00.15s - 47M | 00.31s - 71M
gcc | 00.05s - 20M | 00.07s - 23M | 00.08s - 26M | 00.14s - 36M | 00.30s - 66M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.09s - 39M | 00.13s - 43M | 00.20s - 51M
gcc | 00.05s - 21M | 00.06s - 22M | 00.08s - 25M | 00.13s - 30M | 00.21s - 41M

---

```cpp
using list = mk_list<SIZE, SIZE>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.10s - 40M | 00.14s - 46M | 00.29s - 66M
gcc | 00.05s - 21M | 00.06s - 22M | 00.08s - 26M | 00.13s - 34M | 00.27s - 61M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.08s - 37M | 00.09s - 38M | 00.12s - 41M | 00.18s - 47M
gcc | 00.05s - 21M | 00.07s - 22M | 00.08s - 24M | 00.12s - 28M | 00.20s - 37M